fix(rc,ha,led,ntp,mqtt,storage): roadmap batch R9/R17/R19/R22/R23/R25
R9: replace blocking HAL::delay(100) with non-blocking flag+timer reboot
    pattern in RemoteConsole (constitution Principle X)
R17: remove dead isValidTopic() declaration from MQTT.h + docs
R19: remove unused retryDelayMs field from NTPConfig + tests + docs
R22: remove unused effectDirection field from LEDState + docs
R23: remove dead #if DOMOTICSCORE_WEBUI_ENABLED block from Storage.h
R25: convert enum AlarmFeature to enum class with inline constexpr
     bitwise operator overloads

+++ DomoticsCore-HomeAssistant/include/DomoticsCore/HomeAssistant.h
@@ -251,7 +251,7 @@ public:
         const String& id, const String& name,
         const std::function<void(const String& command, const String& code)>& commandCallback,
         const String& icon = "mdi:shield-home",
-        uint8_t features = AlarmFeature::ArmAway,
+        uint8_t features = static_cast<uint8_t>(AlarmFeature::ArmAway),
         const String& code = "",
         bool codeArmRequired = false,
         bool codeDisarmRequired = false,
